Divide sharing by barcode to other function object.
